fixing httpcontextaccessor
