+++ src/DotNETDevOps.Extensions.AzureFunctions/Abstractions/IWebHostBuilderExtension.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+
+namespace DotNETDevOps.Extensions.AzureFunctions
+{
+    public interface IWebHostBuilderExtension
+    {
+
+        void ConfigureWebHostBuilder(WebHostBuilder builder);
+    }
+}
