+++ src/DotNETDevOps.Extensions.AzureFunctions/Attributes/WebHostBuilderAttribute.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace DotNETDevOps.Extensions.AzureFunctions
+{
+    public class WebHostBuilderAttribute : Attribute
+    {
+        public WebHostBuilderAttribute(Type type)
+        {
+            Type = type;
+        }
+
+        public Type Type { get; }
+    }
+}
